% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.simCIs.R
\name{anova.simCIs}
\alias{anova.simCIs}
\title{Simultaneous CIs Calculation using Tukey Method in One-Way ANOVA}
\usage{
\method{anova}{simCIs}(data, conflev = 0.95)
}
\arguments{
\item{data}{Input data in data dump format, for example,
data = list(x1=c(5, 4, 6), x2=c(7, 8, 4), x3 = c(9, 10, 15, 8))
where xi is the ith sample in one-way anova analysis.}

\item{conflev}{Confidence interval level, e.g., 0.95}
}
\value{
An object with values including
\itemize{
\item  mu.diff: a matrix of pairwise sample mean difference
\item  CI: simultaneous pairwise CI for mu_i - mu_j, which is CI\link{i,j,}.
}
}
\description{
This function calculates the pairwise simultaneous CIs using Tukey Method
in One-Way ANOVA
}
\examples{
data = list(x1=c(5, 4, 6), x2=c(7, 8, 4), x3 = c(9, 10, 15, 8))
anova.simCIs(data, conflev=0.95)

}
