% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_MSE.R
\name{anova_MSE}
\alias{anova_MSE}
\title{MSE Calculation in One-Way ANOVA}
\usage{
anova_MSE(data)
}
\arguments{
\item{data}{Input data in data dump format, for example,
data = list(x1=c(5, 4, 6), x2=c(7, 8, 4), x3 = c(9, 10, 15, 8))
where xi is the ith sample in one-way anova analysis.}
}
\value{
An object with values including
\itemize{
\item  n: sample size in each sample
\item  N: total sample size
\item  k: number of samples
\item  mu: sample mean in each sample
\item  s: sample standard deviation in each sample
\item  SSE: sum of squared error
\item  MSE: mean squared error
}
}
\description{
This function calculates the SSE and MSE in one-way ANOVA.
}
\examples{
data = list(x1=c(5, 4, 6), x2=c(7, 8, 4), x3 = c(9, 10, 15, 8))
anova_MSE(data)

}
