% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.two.p.R
\name{chisq.two.p}
\alias{chisq.two.p}
\title{Chi-square Test for Comparing Two Proportions}
\usage{
chisq.two.p(r1, n1, r2, n2, correct = TRUE, alpha = 0.05, conflev = 0.95, ...)
}
\arguments{
\item{r1}{Number of successes in group 1}

\item{n1}{Number of trials in group 1}

\item{r2}{Number of successes in group 2}

\item{n2}{Number of trials in group 2}

\item{correct}{Continuity correction, TRUE or FALSE}

\item{alpha}{Significance level, default 0.05}

\item{conflev}{Level of confidence interval, default 0.95}

\item{...}{Additional parameters to pass to chisq.test(). Check ?chisq.test for details.}
}
\value{
An object that has components
\describe{
\item{p}{P value}
\item{chisq}{Value of test statistic chi-square}
\item{CV}{Critical value}
\item{conclusion}{Statistical conclusion}
\item{df}{Degrees of freedom for chi-square test}
\item{method}{Test method}
\item{observed}{2 by 2 table for observed values}
\item{expected}{2 by 2 table for expected values}
\item{residuals}{2 by 2 table for Pearson residuals, (observed - expected) / sqrt(expected)}
\item{$standardized.residuals}{2 by 2 table for standardized residuals,
(observed - expected) / sqrt(V), where V is the residual cell variance
(Agresti, 2007, section 2.4.5 for the case where x is a matrix, np(1 - p) otherwise).}
\item{phat}{Estimates of proportions in two groups}
\item{p1.exact.ci}{Exact CI by Clopper-Pearson method for p1}
\item{p2.exact.ci}{Exact CI by Clopper-Pearson method for p2}
\item{p1.ci}{CI for p1 by asymptotic normal method with or without continuity correction according to parameter correct}
\item{p2.ci}{CI for p2 by asymptotic normal method with or without continuity correction according to parameter correct}
}
}
\description{
This function calculates the Chi-square test. Because the test statistic is z value squared,
it is always right-sided, however it is equivalent to two-sided test in Z test.
It is equivalent to function chisq.test(), however provides the convenience of direct
calculation based on number of successes and number of trials in each group.
}
\examples{
#Example: 107 responses out of 245 subjects vs 52 responses out of 232 subjects
chisq.two.p(r1=107, n1=245, r2=52, n2=232, correct=TRUE, alpha=0.05, conflev = 0.95)

chisq.two.p(r1=6, n1=245, r2=38, n2=232, correct=TRUE, alpha=0.05, conflev = 0.95)

}
