% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.mu.diff.R
\name{ci.mu.diff}
\alias{ci.mu.diff}
\title{Confidence Interval (Two-Sided) of mean difference with known or unknown variance}
\usage{
ci.mu.diff(
  x1 = NULL,
  x1bar,
  n1,
  s1,
  sd1 = NULL,
  x2 = NULL,
  x2bar,
  n2,
  s2,
  sd2 = NULL,
  known.variance = FALSE,
  equal.variance = TRUE,
  conflev = 0.95,
  uneqvar.method = "Welch"
)
}
\arguments{
\item{x1}{Sample 1. Default NULL.}

\item{x1bar}{Sample 1 mean}

\item{n1}{Sample 1 size}

\item{s1}{sample 1 standard deviation.}

\item{sd1}{Known standard deviation for population 1.}

\item{x2}{Sample 2. Default NULL.}

\item{x2bar}{Sample 2 mean}

\item{n2}{Sample 2 size}

\item{s2}{sample 2 standard deviation.}

\item{sd2}{Known standard deviation for population 2.}

\item{known.variance}{TRUE or FALSE. Default FALSE, i.e. t distribution.}

\item{equal.variance}{TRUE or FALSE. Default TRUE, i.e. pooled t}

\item{conflev}{Two-sided confidence level}

\item{uneqvar.method}{"Welch" or "Cochran". This option is applicable only
in the scenario of unequal and unknown variances. Default "Welch".}
}
\description{
This function calculates the CI of mean difference with known or unknown variance assuming normal distribution.
}
\examples{
#Example (1). The sample x is provided.
x1 = rnorm(100, mean=10, sd=2)
x2 = rnorm(100, mean=12, sd=2.2)

ci.mu.diff(x1=x1, x2=x2, conflev=0.95, known.variance=FALSE, equal.variance=TRUE)
ci.mu.diff(x1=x1, x2=x2, conflev=0.95, known.variance=FALSE, equal.variance=FALSE)

ci.mu.diff(x1=x1, x2=x2, sd1=2, sd2=2, conflev=0.95, known.variance=TRUE, equal.variance=TRUE)
ci.mu.diff(x1=x1, x2=x2, sd1=2, sd2=2, conflev=0.95, known.variance=TRUE, equal.variance=FALSE)

#Example (2). Sample mean and n are provided.

ci.mu.diff(x1bar=16, n1=20, s1 = 5, x2bar=12, n2=20, s2 = 4.5, conflev=0.95,
known.variance=FALSE, equal.variance=TRUE)

ci.mu.diff(x1bar=16, n1=20, s1 = 5, x2bar=12, n2=20, s2 = 4.5, conflev=0.95,
known.variance=FALSE, equal.variance=FALSE)

ci.mu.diff(x1bar=16, n1=20, sd1 = 5, x2bar=12, n2=20, sd2 = 5, conflev=0.95,
known.variance=TRUE, equal.variance=TRUE)

ci.mu.diff(x1bar=16, n1=20, sd1 = 5, x2bar=12, n2=20, sd2 = 4.5, conflev=0.95,
known.variance=TRUE, equal.variance=FALSE)

ci.mu.diff(x1bar=16, n1=20, s1 = 5, x2bar=12, n2=20, s2 = 4.5, conflev=0.95,
known.variance=FALSE, equal.variance=FALSE, uneqvar.method="Cochran")

ci.mu.diff(x1bar=16, n1=20, s1 = 5, x2bar=12, n2=20, s2 = 4.5, conflev=0.95,
known.variance=FALSE, equal.variance=FALSE, uneqvar.method="Welch")

}
