% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.mu.R
\name{ci.mu}
\alias{ci.mu}
\title{Confidence Interval (Two-Sided) of mean with known or unknown variance}
\usage{
ci.mu(
  xbar,
  n,
  x = NULL,
  s = NULL,
  sd = 1,
  conflev = 0.95,
  method = "Known Variance"
)
}
\arguments{
\item{xbar}{Sample mean}

\item{n}{Sample size}

\item{x}{Sample. Default NULL. When x is available, then use x to calculate xbar, n, and s.}

\item{s}{sample standard deviation Either x or s is required when method = "Unknown Variance"}

\item{sd}{Known standard deviation. Either x or sd is required when method = "Known Variance". This is the population sd}

\item{conflev}{Two-sided confidence level}
}
\description{
This function calculates the CI of mean with known or unknown variance assuming normal distribution.
}
\examples{
#Example (1). The sample x is provided.
x = rnorm(100, mean=10, sd=2)

ci.mu(x=x, conflev=0.95, sd = 2, method="Known Variance")

ci.mu(x=x, conflev=0.95, method="Unknown Variance")

#Example (2). Sample mean and n are provided.

xbar = mean(x); s = sd(x); n = 100

ci.mu(xbar=xbar, sd = 2, n = 100, conflev=0.95, method="Known Variance")

ci.mu(xbar=xbar, s=s, n = 100, conflev=0.95, method="Unknown Variance")

}
