% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.p.diff.R
\name{ci.p.diff}
\alias{ci.p.diff}
\title{Confidence Interval (Two-Sided) of two proportions using normal approximation}
\usage{
ci.p.diff(r1, n1, r2, n2, x1 = NULL, x2 = NULL, conflev = 0.95)
}
\arguments{
\item{r1}{Number of successes in sample 1}

\item{n1}{Sample 1 size}

\item{r2}{Number of successes in sample 2.}

\item{n2}{Sample 2 size}

\item{x1}{Sample 1. When available, re-calculate r1 and n1}

\item{x2}{Sample 2. When available, re-calculate r2 and n2}

\item{conflev}{Two-sided confidence level}
}
\description{
This function calculates the CI of two proportions using normal approximation
}
\examples{
#Example (1). x1 and x2 provided
x1 = sample(x=c(rep(1,30), rep(0,70)))
x2 = sample(x=c(rep(1,40), rep(0,60)))
ci.p.diff(x1=x1, x2=x2, conflev=0.95)

#Example (2). x1 and x2 not provided

ci.p.diff(r1 = 30, n1=100, r2=40, n2=100, conflev=0.95)

}
