% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.test.R
\name{correlation.test}
\alias{correlation.test}
\title{Perform Hypothesis Test for Linear Correlation}
\usage{
correlation.test(x, y, alpha = 0.05, tail = "two")
}
\arguments{
\item{x}{predictor variable}

\item{y}{response variable}

\item{alpha}{Significance level}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}
}
\value{
Linear correlation, CI of linear correlation,
cv for testing the linear correlation, observed value of t-test statistic,
p value.
}
\description{
This functions performs the hypothesis test for linear correlation between x and y.
}
\examples{

x = rnorm(100, mean=40, sd=10)
y = rnorm(100, mean=20, sd=20)
correlation.test (x, y, alpha=0.05, tail="two")

}
