% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cucconi.test.R
\name{cucconi.test}
\alias{cucconi.test}
\title{Cucconi Test comparing two distribution functions}
\usage{
cucconi.test(x, y, method = c("permutation", "bootstrap"))
}
\arguments{
\item{x}{sample 1}

\item{y}{sample 2}

\item{method}{"permutation" or "bootstrap" method to construct the empirical distributions}
}
\value{
p value, test statistic (C)
}
\description{
This function performs Cucconi Test comparing two distribution functions equality.
Implementation of the Cucconi test for the two-sample location-scale problem.
A permutation/bootstrap distribution of the test statistic (C) under the
null hypothesis is used to calculate the p-value.
Reference: Marozzi (2013), p. 1302-1303
}
\examples{
#Example

x <- rnorm(30)
y <- rnorm(30) * 1.3 + 0.5
cucconi.test(x = x, y = y)
cucconi.test(x = x, y = y, method="bootstrap")

}
