% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.bell.shape.R
\name{draw.bell.shape}
\alias{draw.bell.shape}
\title{Draw Bell-Shaped Distribution}
\usage{
draw.bell.shape(peak.loc = 0, width = 1, n = 1000, ...)
}
\arguments{
\item{width}{Spread of each peak}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}

\item{peaks.loc}{Peak location}
}
\description{
This function draws a Bell-Shaped distribution for illustration of statistical distributions.
}
\examples{
#Example.
draw.bell.shape(peak.loc=10, width=2, main="Bell Shaped Distribution")

}
