% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.dist.xbar.normal.R
\name{draw.dist.xbar.normal}
\alias{draw.dist.xbar.normal}
\title{Illustration of sampling distribution of x_bar with a sample drawn from a normal distribution}
\usage{
draw.dist.xbar.normal(n = 100, sigma = 2, mu = 10, ...)
}
\arguments{
\item{n}{Sample size}

\item{sigma}{Standard deviation from a population}

\item{mu}{Mean of a population}

\item{...}{Further graphical parameters}
}
\description{
This function draws the density curves of x and x_bar.
}
\examples{
#Example.
draw.dist.xbar.normal(n=10, sigma=2, mu=10)

}
