% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.multi.modes.R
\name{draw.multi.modes}
\alias{draw.multi.modes}
\title{Draw Multimodal Distribution}
\usage{
draw.multi.modes(
  peaks.loc = c(3, 7, 12),
  width = c(1, 1, 1),
  n = c(1000, 1000, 1000),
  ...
)
}
\arguments{
\item{peaks.loc}{Peak location}

\item{width}{Spread of each peak}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}
}
\description{
This function draws a multimodal distribution for illustration of statistical distributions.
}
\examples{
#Example (1). Unimodal
draw.multi.modes(peaks.loc = c(5),  width=c(3),
xlim=c(-5, 15), ylim=c(0, 0.42), main="Peak: 5, width: 3")

#Example (2). bimodal
draw.multi.modes(peaks.loc = c(5, 15),  width=c(1,3),
xlim=c(-5, 25), ylim=c(0, 0.25),
main="Peak: 5 15, width: 1 3")

#Example (3). 3-modal distribution
draw.multi.modes(peaks.loc = c(5, 15, 25),  width=c(1,2,3),
xlim=c(-5, 35), ylim=c(0, 0.15),
main="Peak: 5 15 25, width: 1 2 3")

}
