% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.normal.dist.R
\name{draw.normal.dist}
\alias{draw.normal.dist}
\title{Plot F distribution density to illustrate the features of F distribution}
\usage{
draw.normal.dist(mu, sigma, ...)
}
\arguments{
\item{mu}{mean parameter in normal distribution}

\item{sigma}{sd parameter in normal distribution}

\item{...}{additional graphic parameters}
}
\value{
graph
}
\description{
This function plots the F distribution density to illustrate the features.
}
\examples{
#N(10, sd=3)
draw.normal.dist(mu=10, sigma=3, xlim=c(-20, 40), ylim=c(0,0.13))

}
