% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.reverse.Jshape.R
\name{draw.reverse.Jshape}
\alias{draw.reverse.Jshape}
\title{Draw Reverse-J Shaped Distribution}
\usage{
draw.reverse.Jshape(median.loc = 10, n = 1000, ...)
}
\arguments{
\item{median.loc}{Location of median}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}
}
\description{
This function draws a reverse-J shaped distribution for illustration of statistical distributions.
}
\examples{
#Example.
draw.reverse.Jshape (median.loc=10)

}
