% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.shade.under.density.R
\name{draw.shade.under.density}
\alias{draw.shade.under.density}
\title{Plot Shade Under a Density Curve}
\usage{
draw.shade.under.density(
  dfun = dnorm,
  range = c(-3, 3),
  shade.interval = c(-1, 1),
  curve.col = "aquamarine4",
  shade.col = "turquoise",
  xlab = "x",
  ylab = "y",
  main = "Density of N(0, 1)",
  ...
)
}
\arguments{
\item{dfun}{density function}

\item{range}{range of x}

\item{shade.interval}{interval for drawing shade under the curve}

\item{curve.col}{color of the density curve}

\item{shade.col}{color of the shade}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label description}

\item{main}{title of the graph#'}

\item{...}{additional parameters pass to dfun() function}
}
\value{
graph
}
\description{
This function plots the F distribution density to illustrate the features.
}
\examples{
#N(0, 1) with shade between (-1, 1)
draw.shade.under.density(dfun = dnorm, range=c(-3, 3), 
shade.interval=c(-1, 1), curve.col = "aquamarine4", shade.col = "turquoise",
xlab="x", ylab="Density Curve of N(0, 1)", mean=0, sd=1)

}
