% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.skewed.shape.R
\name{draw.skewed.shape}
\alias{draw.skewed.shape}
\title{Draw Skewed Distribution}
\usage{
draw.skewed.shape(peak.loc = 10, width = 1, skew = "right", n = 1000, ...)
}
\arguments{
\item{width}{Spread of each peak}

\item{skew}{Optionsl include "right", "left", "no"}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}

\item{peaks.loc}{Peak location}
}
\description{
This function draws a skewed distribution for illustration of statistical distributions.
}
\examples{

#Example.
draw.skewed.shape(peak.loc=10, width=1, skew = "right",
main="Right Skewed Distribution",
xlim=c(0, 50), ylim=c(0, 0.22))

draw.skewed.shape(peak.loc=10, width=2, skew = "right",
main="Right Skewed Distribution",
xlim=c(0, 50), ylim=c(0, 0.22))

draw.skewed.shape(peak.loc=10, width=2, skew = "left",
main="Left Skewed Distribution",
xlim=c(-20, 20), ylim=c(0, 0.17))

draw.skewed.shape(peak.loc=10, width=3, skew = "left",
main="Left Skewed Distribution",
xlim=c(-20, 20), ylim=c(0, 0.17))

draw.skewed.shape(peak.loc=10, width=2, skew = "no",
main="Not Skewed (symmetric) Distribution",
xlim=c(0, 20), ylim=c(0, 0.22))

draw.skewed.shape(peak.loc=10, width=3, skew = "no",
main="Not Skew (symmetric) Distribution",
xlim=c(0, 20), ylim=c(0, 0.22))

}
