% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.triangle.shape.R
\name{draw.triangle.shape}
\alias{draw.triangle.shape}
\title{Draw Triangular Shape Distribution}
\usage{
draw.triangle.shape(peak.loc = 0, slope = 1, n = 10, ...)
}
\arguments{
\item{slope}{Slope of peak}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}

\item{peaks.loc}{Peak location}
}
\description{
This function draws a Triangular Shape distribution for illustration of statistical distributions.
}
\examples{

#Example.
draw.triangle.shape(peak.loc=10, slope=0.25, n=20,
main="Triangle Shaped Distribution",
ylim=c(0, 0.5), xlim=c(7, 13))

draw.triangle.shape(peak.loc=10, slope=1, n=20,
main="Triangle Shaped Distribution",
ylim=c(0, 0.5), xlim=c(7, 13))

draw.triangle.shape(peak.loc=10, slope=4, n=20,
main="Triangle Shaped Distribution",
ylim=c(0, 0.5), xlim=c(7, 13))

}
