% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.uniform.shape.R
\name{draw.uniform.shape}
\alias{draw.uniform.shape}
\title{Draw Uniform Shape Distribution}
\usage{
draw.uniform.shape(range = c(0, 1), n = 1000, ...)
}
\arguments{
\item{range}{Range}

\item{n}{Sample size}

\item{...}{Optional graphical parameters}
}
\description{
This function draws a uniform Shape distribution for illustration of statistical distributions.
}
\examples{

#Example.
draw.uniform.shape(range = c(0, 1), main="Uniform Distribution")

}
