% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.US.households.R
\name{draw.US.households}
\alias{draw.US.households}
\title{Draw Sample Distribution of U.S. Household Size}
\usage{
draw.US.households(
  size = 100,
  p = c(0.27, 0.34, 0.15, 0.14, 0.06, 0.03, 0.01),
  ...
)
}
\arguments{
\item{size}{Number of households to draw}

\item{p}{Population probabilities of household sizes 1, 2, ..., 7+.}

\item{...}{Optional graphical parameters}
}
\description{
This function draws the sample distribution of U.S. household size for illustration of statistical distributions.
}
\examples{
#Example.
#Population Distribution
p = c(0.27, 0.34, 0.15, 0.14, 0.06, 0.03, 0.01)
barplot(p, col="turquoise", space=0,
names.arg = c("1", "2", "3", "4", "5","6","7+"), ylim=c(0, 0.4),
main="Population Distribution (All US Households)", xlab = "Household Size")

#Sample Distribution
draw.US.households(size=100, main="Sample 1 Distribution (100 US Households)")
draw.US.households(size=100, main="Sample 2 Distribution (100 US Households)")
draw.US.households(size=100, main="Sample 3 Distribution (100 US Households)")
draw.US.households(size=100, main="Sample 4 Distribution (100 US Households)")
draw.US.households(size=100, main="Sample 5 Distribution (100 US Households)")
draw.US.households(size=100, main="Sample 6 Distribution (100 US Households)")
draw.US.households(size=1000, main="Sample 7 Distribution (1000 US Households)")
draw.US.households(size=100000, main="Sample 8 Distribution (100,000 US Households)")

}
