% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical.boot.lm.R
\name{empirical.boot.lm}
\alias{empirical.boot.lm}
\title{Bootstrap linear regression - Empirical method}
\usage{
empirical.boot.lm(x, y, conflev = 0.95, seed = 2023, B = 10000)
}
\arguments{
\item{x}{covariate}

\item{y}{response}

\item{conflev}{confidence level for producing the CI of beta1. Default 0.95.}

\item{B}{number of bootstrapping replicates}
}
\value{
An object with values including
\itemize{
\item  beta: coefficients
\item  beta0.CI: CI of beta0
\item  beta1.CI: CI of beta1
\item  se.beta: se of beta0 and beta1
}
}
\description{
This function performs Bootstrap linear regression - Empirical method
}
\examples{
#Example
set.seed(2023)
x <- rnorm(50, mean=10, sd=2)
y <- rnorm(50, mean=2*x+3, sd=2)
empirical.boot.lm(x, y)

}
