% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{ggplot_it}
\alias{ggplot_it}
\title{Plots a normal histogram}
\usage{
ggplot_it(n = 100)
}
\arguments{
\item{n}{Number of random draws.}
}
\value{
Plots a histogram of n standard normals.
}
\description{
Plots a normal histogram
}
\examples{
ggplot_it()
ggplot_it(1000)
}

