% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.error.mu.R
\name{margin.error.mu}
\alias{margin.error.mu}
\title{Margin of Error for Population Mean}
\usage{
margin.error.mu(
  x = NULL,
  conflev = 0.95,
  s,
  sigma = NULL,
  n,
  method = "Unknown Variance"
)
}
\arguments{
\item{x}{sample}

\item{conflev}{Confidence level}

\item{s}{sample standard deviation}

\item{sigma}{population standard deviation}

\item{n}{sample size}

\item{method}{"Known Variance" or "Unknown Variance"}
}
\value{
E
}
\description{
This function calculates the margin of error for population mean
}
\examples{
#Example

x = rnorm(30)
margin.error.mu(x=x, sigma=1, conflev=0.95, method="Known Variance")
margin.error.mu(n=length(x), sigma=1, conflev=0.95, method="Known Variance")
margin.error.mu(x=x, conflev=0.95, method="Unknown Variance")
margin.error.mu(s=sd(x), n=length(x), conflev=0.95, method="Unknown Variance")


}
