% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.p.CI.R
\name{n.p.CI}
\alias{n.p.CI}
\title{Sample Size calculation for achieving adequate precision in population proportion estimate}
\usage{
n.p.CI(E = 0.02, conflev = 0.95, g = c(0.4, 0.7))
}
\arguments{
\item{E}{Precision of proportion estimate, i.e. half of CI length}

\item{conflev}{Confidence level, eg, 0.95}

\item{g}{A range of likely population proportion}
}
\description{
This function calculates the sample size required to achieve adequate precision in population proportion estimate
}
\examples{
Example
n.p.CI(E=0.1, conflev=0.95, g=c(0.1, 0.3))

}
