% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLR.R
\name{OLR}
\alias{OLR}
\title{Linear Regression Using OLR Method}
\usage{
OLR(x, y, conflev = 0.95, new.x = NULL)
}
\arguments{
\item{x}{covariate}

\item{y}{response}

\item{conflev}{confidence level for producing the CI of beta1. Default 0.95.}

\item{new.x}{a value of x for prediction of y. Default NULL.}
}
\value{
An object having values including
\itemize{
\item  Sxx: XX sum of square sum_(x_i-xbar)^2
\item  Sxy: XY sum of square sum_(x_i-xbar)(y_i-ybar)
\item  beta0: coefficient for intercept; beta0=ybar-beta1*xbar
\item  beta1: coefficient for x; beta1 = Sxy/Sxx
\item  SST: Total sum of squares
\item  SSR: Regression sum of squares
\item  SSE: Error sum of squares
\item  r2:  Coefficient of determination, r2 = SSR / SST
\item  r2adj:  Adjusted coefficient of determination, 1-(SSE / (n-k))/(SST/(n-1)) = 1 - MSE/MST
\item  anova: ANOVA table
\item  sigma.hat residual standard error, sqrt(MSE)
\item  se.beta1 se(beta1)
\item  t0: t-test statistic
\item  p: p value from the t-test
\item F0: F value from the F-test for the covariate x
\item p.F.test: p value from the F test (df1=1, df2=n-2).
\item CI: Confidence interval of beta1. Default 95\\% CI
\item new.y: Predicted value of y given x = new.x
\item CI.new.x: Confidence interval of population mean of new.y when x = new.x
\item predCI.new.x: Prediction interval of new.y when x = new.x
}
}
\description{
This function fits the data for linear regression
}
\examples{
#Example
set.seed(2023)
x <- rnorm(50, mean=10, sd=2)
y <- rnorm(50, mean=2*x+3, sd=2)
OLR(x, y)

OLR(x, y, new.x=20)

}
