% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.mean.perm.R
\name{one.mean.perm}
\alias{one.mean.perm}
\title{Permutation Test of One Population Mean}
\usage{
one.mean.perm(
  x,
  mu0,
  tail = "two",
  n.perm = 5000,
  save = FALSE,
  seed = 2023,
  conflev = 0.95
)
}
\arguments{
\item{x}{sample data}

\item{mu0}{mean under H0}

\item{tail}{type of test: "two", "left", "right"}

\item{n.perm}{number of permutations}

\item{save}{TRUE or FALSE to save test statistics from permutations}

\item{seed}{seed}

\item{conflev}{Confidence level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval by permutation method
\item  p: p value
\item  T: Test statistic
\item  mean: Sample mean
}
}
\description{
This function performs the permutation test of one population mean
}
\examples{
x <- rnorm(10, mean=10, sd=3)
one.mean.perm(x=x, mu0=10, tail="two")
one.median.sign(x=x, m0=10, tail="two")

}
