% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.mean.z.R
\name{one.mean.z}
\alias{one.mean.z}
\title{One Mean z test}
\usage{
one.mean.z(
  x,
  xbar = NULL,
  n = NULL,
  mu0 = 0,
  sigma = 1,
  tail = "two",
  alpha = 0.05
)
}
\arguments{
\item{x}{A sample data set of observations}

\item{xbar}{Sample mean. Default NULL. If xbar is available, then x is ignored.}

\item{n}{Sample size. Default NULL.}

\item{mu0}{Population mean under H0}

\item{sigma}{Population standard deviation}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{alpha}{Significance level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100
\item  xbar: Sample mean
\item  n: Sample size
\item  margin: Margin of CI
\item  z0: Value of z test statistic under H0
\item  p: p value
\item  CV: Critical value
\item  conclusion: Testing result
}
}
\description{
This function performs one-mean Z test assuming population variance known.
}
\examples{
#Example.
x <- c( 57.3, 57.5, 59.0, 56.5, 61.3,
        57.6, 59.2, 65.0, 60.1, 59.7,
        62.6, 52.6, 60.7, 62.3, 65.2,
        54.8, 55.4, 55.5, 57.8, 58.7,
        57.8, 60.9, 75.3, 60.6, 58.1,
        55.9, 61.6, 59.6, 59.8, 63.4,
        54.7, 60.2, 52.4, 58.3, 66.0)
one.mean.z(x=x, mu0=60, sigma=3.2, tail="two", alpha=0.05)

#Use xbar
one.mean.z(xbar=mean(x), n = length(x), mu0=60, sigma=3.2, tail="two", alpha=0.05)

}
