% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.median.sign.R
\name{one.median.sign}
\alias{one.median.sign}
\title{Sign Test of One Population Median}
\usage{
one.median.sign(x, m0, tail = "two")
}
\arguments{
\item{x}{sample data}

\item{m0}{median under H0}

\item{tail}{type of test: "two", "left", "right"}
}
\value{
p value
}
\description{
This function performs the sign test of one population median.
}
\examples{
x <- rnorm(10, mean=10, sd=3)
one.median.sign(x=x, m0=10, tail="two")
#compared to one sample t-test
one.mean.t(x=x, mu0=10, tail="two")

}
