% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.sd.chisq.test.R
\name{one.sd.chisq.test}
\alias{one.sd.chisq.test}
\title{One sd chi-square test}
\usage{
one.sd.chisq.test(s, n, alpha = 0.05, sd0, tail = "left", conflev = 0.95)
}
\arguments{
\item{s}{Sample sd.}

\item{n}{Sample size.}

\item{alpha}{Significance level}

\item{sd0}{Population sd under H0}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{conflev}{confidence interval level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100
\item  chisq0: Value of chi-square test statistic under H0
\item  p: p value
\item  CV: Critical value
\item  conclusion: Testing result
}
}
\description{
This function performs one-sd chi-square test
}
\examples{
x <- c(120.94, 118.58, 119.41, 120.23, 121.13, 118.22, 119.71, 121.09, 120.56, 119.11)
one.sd.chisq.test(s=sd(x), n=length(x), alpha=0.05, sd0=4, tail="two", conflev = 0.95)


}
