% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_normals}
\alias{plot_normals}
\title{Plots normal densities}
\usage{
plot_normals(n_dens = 3)
}
\arguments{
\item{n_dens}{Number of densities to draw with standard deviations from 1 to n_dens}
}
\value{
ggplot2 plot object
}
\description{
Plots normal densities
}
\examples{
plot_normals()
plot_normals(5)
}

