% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_sampling_dist}
\alias{plot_sampling_dist}
\title{Plots the distributions of the sample means from the population.}
\usage{
plot_sampling_dist(pop, n = 1e+05, size = c(2, 5, 30), facets = FALSE)
}
\arguments{
\item{pop}{Vector containing the population (anything over 10,000 will take a long time to run on most ocmputers)}

\item{n}{Number of replicas of each sample}

\item{size}{Vector containing sizes for each sampling distribution}

\item{facets}{Set to TRUE if you want plot each distribution on a separate facet}
}
\value{
ggplot2 plot object
}
\description{
Plots the distributions of the sample means from the population.
}
\examples{
x <- runif(10000)
plot_sampling_dist(x, facets = TRUE)
plot_sampling_dist(x, n = 1e6, size = c(2, 10, 50, 100))
}

