% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.binom.R
\name{pois.binom}
\alias{pois.binom}
\title{Experiment to use Poisson (lambda = np) to approximate Binomial(n, p)}
\usage{
pois.binom(n, p, x.range = NULL, out = FALSE)
}
\arguments{
\item{n}{Size of sample for binomial distribution (n, p)}

\item{p}{Probability of success for binomial distribution (n, p)}
}
\value{
Probabilities of two distributions
}
\description{
This function demonstrates the approximation of Binomial(n, p) by Poisson (lambda = np)
}
\examples{
#Binom(100, 0.3)
pois.binom(n = 100, p = 0.3)

pois.binom(n = 1000, p = 0.003)

}
