% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.sd.R
\name{pop.sd}
\alias{pop.sd}
\title{Population variance for a finite population}
\usage{
pop.sd(x)
}
\arguments{
\item{x}{A numeric array that consists of all possible observations in an entire finite population}
}
\description{
This function calculates the population variance for a finite population.
}
\examples{
#Example.
pop.sd(c(1,2,2,3))


}
