% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.z.test.R
\name{power.z.test}
\alias{power.z.test}
\title{Sample size calculation for One sample z test}
\usage{
power.z.test(
  power = 0.9,
  delta = 1,
  sd = 1,
  alternative = "one.sided",
  sig.level = 0.05,
  type = "one.sample"
)
}
\arguments{
\item{power}{Power}

\item{delta}{Mean difference: mu - mu0}

\item{sd}{Population sd}

\item{alternative}{"one.sided"}

\item{sig.level}{significance level, alpha}

\item{type}{"one.sample"}
}
\description{
This function calculates the sample size based on one sample z test
}
\examples{
#Example.
power.z.test(power = 0.9, delta = 1, sd=1, alternative = "one.sided", 
sig.level=0.05, type="one.sample")

}
