% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/introstats.R
\docType{package}
\name{stats250}
\alias{stats250}
\alias{stats250-package}
\title{stats250: A package of Helper Functions for STATS 250, the
Introductory Statistics class at the University of Michigan.}
\description{
stats250: A package of Helper Functions for STATS 250, the
Introductory Statistics class at the University of Michigan.
}

