% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.sample.R
\name{strat.sample}
\alias{strat.sample}
\title{Stratified sampling with proportional allocation}
\usage{
strat.sample(x, strata, size, proportional = TRUE)
}
\arguments{
\item{x}{an array that represents the entire population}

\item{strata}{stratum indicator}

\item{size}{sample size}

\item{proportional}{TRUE or FALSE. If true, the sample is proportional to the strata size.}
}
\description{
This function sample data with proportional allocation.
}
\examples{
#100 numbers from 1 to 100
#1st stratum size 70 and 2nd stratum size 30
#pick a sample of size 10: 7 from 1st stratum and 3 from 2nd stratum which is proportional.
strat.sample(x=1:100, strata=c(rep(1, 70), rep(2, 30)), size=10, proportional = TRUE)


}
