% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.coin.R
\name{toss.coin}
\alias{toss.coin}
\title{Simulation of tossing a coin}
\usage{
toss.coin(n)
}
\arguments{
\item{n}{Number of tosses}
}
\value{
A data set of outcomes
}
\description{
This function performs the simulation of tossing balanced coins.
}
\examples{
#Toss 5 times
toss.coin(n=5)

}
