% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.mean.boot.R
\name{two.mean.boot}
\alias{two.mean.boot}
\title{Bootstrapping method comparing two population means}
\usage{
two.mean.boot(
  x1,
  x2,
  tail = "two",
  conflev = 0.95,
  n.boot = 10000,
  seed = 2023
)
}
\arguments{
\item{x1}{sample 1}

\item{x2}{sample 2}

\item{tail}{type of test "two", "left", "right". Default "two"}

\item{conflev}{Confidence level}

\item{n.boot}{number of bootstrapping samples, default 10000}

\item{seed}{default 2023}
}
\value{
An object having values including
\itemize{
\item  p: p value
\item  mean_diff: Sample mean difference
\item  mean_diff_CI: confidence interval for mu1 - mu2
}
}
\description{
This function performs the bootstrapping method to compare two population means
}
\examples{
#Example
set.seed(2023)
x1 <- rnorm(50, mean=10, sd=2)
x2 <- rnorm(30, mean=12, sd=2.5)
two.mean.perm(x1, x2, tail="two", conflev=0.95, n.perm=10000, seed=2023)
two.mean.boot(x1, x2, tail="two", n.boot=10000, seed=2023)


}
