% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.mean.t.R
\name{two.mean.t}
\alias{two.mean.t}
\title{Two Mean t test}
\usage{
two.mean.t(
  x1,
  x2,
  x1bar = NULL,
  n1 = NULL,
  s1 = NULL,
  x2bar = NULL,
  n2 = NULL,
  s2 = NULL,
  equal.variance = TRUE,
  uneqvar.method = "Welch",
  tail = "two",
  delta0 = 0,
  alpha = 0.05
)
}
\arguments{
\item{x1}{Sample 1}

\item{x2}{Sample 2}

\item{x1bar}{Sample 1 mean. Default NULL. If x1bar is available, then x1 is ignored.}

\item{n1}{Sample 1 size. Default NULL.}

\item{s1}{Sample 1 sd. Default NULL.}

\item{x2bar}{Sample 2 mean. Default NULL. If x2bar is available, then x2 is ignored.}

\item{n2}{Sample 2 size. Default NULL.}

\item{s2}{Sample 2 sd. Default NULL.}

\item{equal.variance}{TRUE or FALSE}

\item{uneqvar.method}{"Welch" or "Cochran". This option is applicable only
in the scenario of unequal variance}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{delta0}{Population mean difference mu1-mu2 under H0. Default 0}

\item{alpha}{Significance level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100 for mu1-mu2
\item  xbar: Sample means
\item  n: Sample size
\item  s: Sample sd
\item  sp: Pooled sample sd
\item  se: Sample se when using unpooled method
\item  margin: Margin of CI
\item  t0: Value of t test statistic under H0
\item  df: Degree of freedom for t test
\item  p: p value
\item  CV: Critical value
\item  conclusion: Testing result
\item  equal.variance: Variance assumption
\item  uneqvar.method: Method for handling unequal variance in t test
}
}
\description{
This function performs two-mean t test assuming population variances unknown.
}
\examples{
#Example.
x <- c( 57.3, 57.5, 59.0, 56.5, 61.3,
        57.6, 59.2, 65.0, 60.1, 59.7,
        62.6, 52.6, 60.7, 62.3, 65.2,
        54.8, 55.4, 55.5, 57.8, 58.7,
        57.8, 60.9, 75.3, 60.6, 58.1,
        55.9, 61.6, 59.6, 59.8, 63.4,
        54.7, 60.2, 52.4, 58.3, 66.0)
y <- x - 2*rnorm(length(x))

two.mean.t(x1=x, x2=y, s1=sd(x), s2=sd(y), tail="right", alpha=0.05)

#Use summary statistics x1bar, x2bar, n1, s1, n2, s2
two.mean.t(x1bar=59.01, n1=53, s1=44.89,
                     x2bar=46.61, n2=54, s2=34.85,
                     tail="right", alpha=0.05)

two.mean.t(x1bar=59.01, n1=53, s1=44.89,
                     x2bar=46.61, n2=54, s2=34.85,
                     tail="right", alpha=0.05,
                     equal.variance=FALSE, uneqvar.method = "Welch")

two.mean.t(x1bar=59.01, n1=53, s1=44.89,
                     x2bar=46.61, n2=54, s2=34.85,
                     tail="right", alpha=0.05,
                     equal.variance=FALSE, uneqvar.method = "Cochran")

}
