% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.mean.z.R
\name{two.mean.z}
\alias{two.mean.z}
\title{Two Mean z test}
\usage{
two.mean.z(
  x1,
  x2,
  x1bar = NULL,
  n1 = NULL,
  sigma1,
  x2bar = NULL,
  n2 = NULL,
  sigma2,
  tail = "two",
  delta0 = 0,
  alpha = 0.05
)
}
\arguments{
\item{x1}{Sample 1}

\item{x2}{Sample 2}

\item{x1bar}{Sample 1 mean. Default NULL. If x1bar is available, then x1 is ignored.}

\item{n1}{Sample 1 size. Default NULL.}

\item{sigma1}{Population 1 standard deviation. Required}

\item{x2bar}{Sample 2 mean. Default NULL. If x2bar is available, then x2 is ignored.}

\item{n2}{Sample 2 size. Default NULL.}

\item{sigma2}{Population 2 standard deviation. Required}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{delta0}{Population mean difference mu1-mu2 under H0. Default 0}

\item{alpha}{Significance level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100 for mu1-mu2
\item  xbar: Sample means
\item  n: Sample size
\item  margin: Margin of CI
\item  z0: Value of z test statistic under H0
\item  p: p value
\item  CV: Critical value
\item  conclusion: Testing result
}
}
\description{
This function performs two-mean z test assuming population variances known.
}
\examples{
#Example.
x <- c( 57.3, 57.5, 59.0, 56.5, 61.3,
        57.6, 59.2, 65.0, 60.1, 59.7,
        62.6, 52.6, 60.7, 62.3, 65.2,
        54.8, 55.4, 55.5, 57.8, 58.7,
        57.8, 60.9, 75.3, 60.6, 58.1,
        55.9, 61.6, 59.6, 59.8, 63.4,
        54.7, 60.2, 52.4, 58.3, 66.0)
y <- x - 2*rnorm(length(x))

two.mean.z(x1=x, x2=y, sigma1=44.89, sigma2=46.61, tail="right", alpha=0.05)

#Use x1bar, x2bar, n1, n2
two.mean.z(x1bar=mean(x), n1=length(x), sigma1=44.89,
          x2bar=mean(y), n2=length(y), sigma2=46.61,
          tail="right", alpha=0.05)

}
