% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.p.z.R
\name{two.p.z}
\alias{two.p.z}
\title{One sample proportion test}
\usage{
two.p.z(
  x,
  y,
  x1 = NULL,
  n1 = NULL,
  x2 = NULL,
  n2 = NULL,
  tail = "left",
  alpha = 0.05,
  conflev = 0.95
)
}
\arguments{
\item{x}{A binary sample data set of observations and each element having a value of 0 or 1.}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{alpha}{Significance level}

\item{phat}{Sample proportion. Default NULL. If phat is available, then x is ignored.}

\item{n}{Sample size. Default NULL.}

\item{p0}{Population mean under H0, default 0}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100
\item  phat: Sample proportion
\item  n: Sample size
\item  margin: Margin of CI
\item  z0: Value of z test statistic under H0
\item  p: p value
\item  CV: Critical value
\item  se: Standard error
\item  conclusion: Testing result
}
}
\description{
This function performs one sample proportion test using normal approximation.
}
\examples{
#Example.
x <- c(rep(1, 24), rep(0, 301-24))
one.p.z(x=x, p0=0.063, tail="right", alpha=0.05)

#Use summary statistics
two.p.z(x1=21, n1=150, x2=48, n2=200, tail="two", alpha=0.05, conflev=0.95)

}
