% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.sd.F.test.R
\name{two.sd.F.test}
\alias{two.sd.F.test}
\title{Two sd F Test}
\usage{
two.sd.F.test(alpha, s1, n1, s2, n2, tail = "left", conflev = 0.95)
}
\arguments{
\item{alpha}{Significance level}

\item{s1}{Sample 1 sd.}

\item{n1}{Sample size.}

\item{s2}{Sample 2 sd.}

\item{n2}{Sample size.}

\item{tail}{Side of hypothesis testing: "two", "left", "right". Default "two"}

\item{conflev}{confidence interval level}
}
\value{
An object with values including
\itemize{
\item  CI: Confidence interval (1-alpha)*100
\item  F0: Value of F test statistic under H0
\item  p: p value
\item  CV: Critical value
\item  conclusion: Testing result
}
}
\description{
This function performs two sd F test for comparing two population sd.
}
\examples{
two.sd.F.test(alpha=0.05, s1=30.5, n1=8, s2=64.5, n2=15, tail="left", conflev=0.9)


}
