% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_package}
\alias{update_package}
\title{Updates the introstats pacakge from github}
\usage{
update_package()
}
\description{
Updates the introstats pacakge from github
}
\examples{
update_package()
}

