% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.score.R
\name{z.score}
\alias{z.score}
\title{z-score}
\usage{
z.score(x, mu = NULL, sigma = NULL)
}
\arguments{
\item{x}{A numeric array that consists of all possible observations in an entire finite population}
}
\description{
This function calculates the z score.
}
\examples{
#Example.
z.score(c(1,2,2,3))

z.score(c(1,2,2,3), mu=2.5, sigma=1)

}
