% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence Interval (Two-Sided) Calculation}
\usage{
ci(r, n, conflev, method = "Clopper-Pearson Exact")
}
\arguments{
\item{r}{number of successes}

\item{n}{size}

\item{conflev}{Two-sided confidence level}

\item{method}{Confidence interval method. Options include "Clopper-Pearson Exact",
"Wilson Score", "Wilson Score with Continuity", "Asymptotic Normal", and "Asymptotic Normal with Continuity".}
}
\description{
This function implements multiple methods for CI calculation for a binomial success
probability for r successes out of n trials with confidence level.
}
\examples{

ci(r=4, n=20, conflev=0.95, method="Clopper-Pearson Exact")
ci(r=4, n=20, conflev=0.95, method="Wilson Score")
ci(r=4, n=20, conflev=0.95, method="Wilson Score with Continuity")
ci(r=4, n=20, conflev=0.95, method="Asymptotic Normal")
ci(r=4, n=20, conflev=0.95, method="Asymptotic Normal with Continuity")

}
