% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.mu.CI.R
\name{n.mu.CI}
\alias{n.mu.CI}
\title{Sample Size calculation for achieving adequate precision in population mean estimate}
\usage{
n.mu.CI(E, sigma, conflev = 0.95, method = "Z")
}
\arguments{
\item{E}{Precision of mean estimate, i.e. half of CI length}

\item{sigma}{Population standard deviation}

\item{conflev}{Confidence level, eg, 0.95}

\item{method}{"Z" for Z interval or "t" for t interval}
}
\description{
This function calculates the sample size required to achieve adequate precision in population mean estimate
}
\examples{
Example
n.mu.CI(E=0.6, sigma=1, conflev=0.99)

n.mu.CI(E=0.6, sigma=1, conflev=0.99, method="t")

}
