% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.sample.R
\name{strat.sample}
\alias{strat.sample}
\title{Stratified sampling with proportional allocation}
\usage{
strat.sample(x, strata, size, proportional = TRUE)
}
\arguments{
\item{x}{an array that represents the entire population}

\item{strata}{stratum indicator}

\item{size}{sample size}

\item{proportional}{TRUE or FALSE. If true, the sample is proportional to the strata size.}
}
\description{
This function sample data with proportional allocation.
}
\examples{

strat.sample(x=rnorm(100), strata=c(rep(1, 20), rep(2, 40), rep(3, 40)), size=10, proportional = TRUE)
strat.sample(x=rnorm(100), strata=c(rep(1, 20), rep(2, 40), rep(3, 40)), size=12, proportional = TRUE)

}
